Storage controller resource UT fix
